feat: Implement scrolling animation to selected movie
Clicking on a movie now correctly scrolls to the movie which is placed directly above the footer.
Additionally, the footer box has been correctly implemented to place text next to the trailer when there is room or under it otherwise.
Javascript functions were placed in the $(document).ready() function so that they are not run until the document is ready.